create confirm.js, move layout funcs from menu.js for reuse
